+++ R/build-home.R
@@ -126,7 +126,7 @@
 #' `authors`, `dev` (badges); and you can add custom components.
 #' The example below creates a sidebar whose only components will be the
 #' authors section, a custom section, and a Dev Status section if there are
-#' badges.
+#' badges. The `text` field has to be Markdown or raw HTML.
 #'
 #' ```
 #' home:
@@ -135,7 +135,7 @@
 #'     components:
 #'       custom:
 #'         title: Funding
-#'         html: We are grateful for funding!
+#'         text: We are grateful for funding!
 #' ```
 #'
 #' You can provide a ready-made sidebar HTML: